Refactor package structure and fix import errors in PowerUpFactory
