+++ core/src/no/ntnu/game/Factory/Button/AbstractButtonFactory.java
@@ -1,4 +1,4 @@
-package no.ntnu.game.Models.Factory.Button;
+package no.ntnu.game.Factory.Button;
 
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.ui.Button;
