+++ core/src/no/ntnu/game/Factory/Button/RectangleButtonFactory.java
@@ -1,4 +1,4 @@
-package no.ntnu.game.Models.Factory.Button;
+package no.ntnu.game.Factory.Button;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
