+++ core/src/no/ntnu/game/Models/TreeWithPowerUp.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.math.MathUtils;
 
-import no.ntnu.game.Models.Factory.PowerUp.PowerUpFactory;
+import no.ntnu.game.Factory.PowerUp.PowerUpFactory;
 
 import java.util.Objects;
