+++ core/src/no/ntnu/game/Views/CreateOrJoinRoomScreen.java
@@ -10,8 +10,8 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 
 import no.ntnu.game.Controllers.GameRoomController;
 import no.ntnu.game.Controllers.PlayerController;
-import no.ntnu.game.Models.Factory.Button.RectangleButtonFactory;
-import no.ntnu.game.Models.Factory.Textfield.TextFieldFactory;
+import no.ntnu.game.Factory.Button.RectangleButtonFactory;
+import no.ntnu.game.Factory.Textfield.TextFieldFactory;
 
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
