+++ core/src/no/ntnu/game/Views/FastestKnightGameScreen.java
@@ -18,10 +18,10 @@ import java.util.Objects;
 
 import no.ntnu.game.Controllers.GameRoomController;
 import no.ntnu.game.Controllers.KnightController;
+import no.ntnu.game.Factory.Button.CircleButtonFactory;
+import no.ntnu.game.Factory.Button.RectangleButtonFactory;
 import no.ntnu.game.Models.Timer;
 import no.ntnu.game.Models.TreeWithPowerUp;
-import no.ntnu.game.Models.Factory.Button.CircleButtonFactory;
-import no.ntnu.game.Models.Factory.Button.RectangleButtonFactory;
 import no.ntnu.game.firestore.GameRoom;
 
 /**
