+++ core/src/no/ntnu/game/Views/PlayerLoginScreen.java
@@ -9,8 +9,8 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 
 import no.ntnu.game.Controllers.PlayerController;
-import no.ntnu.game.Models.Factory.Button.RectangleButtonFactory;
-import no.ntnu.game.Models.Factory.Textfield.TextFieldFactory;
+import no.ntnu.game.Factory.Button.RectangleButtonFactory;
+import no.ntnu.game.Factory.Textfield.TextFieldFactory;
 import no.ntnu.game.firestore.Player;
 
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
