+++ core/src/no/ntnu/game/Views/SingleOrMultiplayerScreen.java
@@ -13,7 +13,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.Button;
 
 import no.ntnu.game.Controllers.GameRoomController;
 import no.ntnu.game.Controllers.PlayerController;
-import no.ntnu.game.Models.Factory.Button.RectangleButtonFactory;
+import no.ntnu.game.Factory.Button.RectangleButtonFactory;
 
 /**
  * Main Menu View class to render main menu screen
